Add files via upload
Here we made a guessing game. It first asks the user if they want to play a game or not. If they input no, then the program terminates. Once they input yes, then it asks for two number inputs. Once the two numbers are inputted, then the computer will generate a random number between the two inputs, and stores it in a variable. 
Then you enter the while loop where it asks you to guess the random number, if you get it right, yay! If you don't get it right, then it gives you another chance and gives a hint where it tells you if the random number is greater or less than your guess. 
If it takes you more than 10 tries, then it will let you know, and it is not nice!